refactor(Settings): 拆分 Settings.js 并移动相关组件
