+++ src/components/Settings/SettingsModal.js
@@ -1,6 +1,6 @@
 import { Modal } from "@arco-design/web-react";
 
-import { useStore } from "../Store";
+import { useStore } from "../../Store";
 import Settings from "./Settings";
 
 export default function SettingsModal() {
